chore: move ImageSpacer into components/ui folder
